Speed system changed. Shell disable script added. Difficulties added.
Speed is now increased every time a new scene is loaded and has no dependence on scene names, it just has a base speed set in the GameMGMT script (and is incrementing correctly).

New script added to parent shell object to disable the shell UI components and all shells themselves if the player has selected the relaxed mode.

All three difficulties are now selectable and start the game with the correct preferences (no game over functionality is present yet to make editing easier).